Share dates calculation logic
