+++ common.mjs
@@ -65,3 +65,43 @@ export function generateCalendarMatrix(year, month) {
 
   return weeks;
 }
+
+//Get the date (day number) of the nth weekday in a given month.
+
+export function getNthWeekdayOfMonth(year, month, weekdayName, occurrence) {
+  const weekdayNames = [
+    "Monday",
+    "Tuesday",
+    "Wednesday",
+    "Thursday",
+    "Friday",
+    "Saturday",
+    "Sunday",
+  ];
+  const weekday = weekdayNames.indexOf(weekdayName);
+  if (weekday === -1) throw new Error(`Invalid weekday name: ${weekdayName}`);
+
+  const daysInMonth = getDaysInMonth(year, month);
+  const days = [];
+
+  for (let d = 1; d <= daysInMonth; d++) {
+    const date = new Date(year, month - 1, d);
+    const jsWeekday = (date.getDay() + 6) % 7; // Monday=0
+    if (jsWeekday === weekday) days.push(d);
+  }
+
+  switch (occurrence?.toLowerCase()) {
+    case "first":
+      return days[0];
+    case "second":
+      return days[1];
+    case "third":
+      return days[2];
+    case "fourth":
+      return days[3];
+    case "last":
+      return days[days.length - 1];
+    default:
+      return null;
+  }
+}
